added support for other transaction messages

+++ WebPay/Request/Transaction.cs
@@ -7,6 +7,7 @@ using System.Xml.Serialization;
 
 namespace WebPay.Request
 {
+
     [XmlRoot("transaction")]
     public class Transaction
     {
@@ -50,7 +51,7 @@ namespace WebPay.Request
         public string OrderNumber { get; set; }
 
         [XmlElement("amount")]
-        public int Amount { get; set; }
+        public double Amount { get; set; }
 
         [XmlElement("currency")]
         public Currency Currency { get; set; }
